versions
This commit adds a new function savepagenow.get_versions which allows
you to iterate through archived versions for a URL. It uses the Internet
Archive's CDX API to determine them, and will safely page through all the
results. Use the limit parameter to limit the total number URLs that the
iterator returns. You can also control the total number of URLs that are
requested at a time from the CDX API using the chunk_limit parameter.

The CLI exposes this functionality with the --versions option, e.g.

    savepagenow --versions https://www.pastpages.org